requirements.txt
pyTenable==1.2.8
deb-pkg-tools==8.4
.
